Make Rect use i16 rather than f32

+++ src/engine.rs
@@ -110,10 +110,10 @@ impl GameLoop {
 }
 
 pub struct Rect {
-    pub x: f32,
-    pub y: f32,
-    pub width: f32,
-    pub height: f32,
+    pub x: i16,
+    pub y: i16,
+    pub width: i16,
+    pub height: i16,
 }
 
 impl Rect {
@@ -242,8 +242,8 @@ impl Image {
         let bounding_box = Rect {
             x: position.x.into(),
             y: position.y.into(),
-            width: element.width() as f32,
-            height: element.height() as f32,
+            width: element.width() as i16,
+            height: element.height() as i16,
         };
         Self { element, position, bounding_box }
     }
@@ -261,11 +261,11 @@ impl Image {
     }
 
     pub fn set_x(&mut self, x: i16) {
-        self.bounding_box.x = x as f32;
+        self.bounding_box.x = x;
         self.position.x = x;
     }
 
     pub fn right(&self) -> i16 {
-        (self.bounding_box.x + self.bounding_box.width) as i16
+        self.bounding_box.x + self.bounding_box.width
     }
 }
