+++ src/game.rs
@@ -115,10 +115,10 @@ impl Game for WalkTheDog {
     }
     fn draw(&self, renderer: &Renderer) {
         renderer.clear(&Rect {
-            x: 0.0,
-            y: 0.0,
-            width: 600.0,
-            height: 600.0,
+            x: 0,
+            y: 0,
+            width: 600,
+            height: 600,
         });
         if let WalkTheDog::Loaded(walk) = self {
             walk.backgrounds.iter().for_each(|background| {
@@ -166,18 +166,18 @@ impl RedHatBoy {
         let sprite = self.current_sprite().expect("Cell not found.");
         Rect {
             x: (self.state_machine.context().position.x +
-                sprite.sprite_source_size.x as i16).into(),
+                sprite.sprite_source_size.x).into(),
             y: (self.state_machine.context().position.y +
-                sprite.sprite_source_size.y as i16).into(),
+                sprite.sprite_source_size.y).into(),
             width: sprite.frame.w.into(),
             height: sprite.frame.h.into(),
         }
     }
 
     fn bounding_box(&self) -> Rect {
-        const X_OFFSET: f32 = 18.0;
-        const Y_OFFSET: f32 = 14.0;
-        const WIDTH_OFFSET: f32 = 28.0;
+        const X_OFFSET: i16 = 18;
+        const Y_OFFSET: i16 = 14;
+        const WIDTH_OFFSET: i16 = 28;
         let mut bounding_box = self.destination_box();
         bounding_box.x += X_OFFSET;
         bounding_box.width -= WIDTH_OFFSET;
@@ -216,7 +216,7 @@ impl RedHatBoy {
         self.state_machine = self.state_machine.transition(Event::Jump);
     }
 
-    fn land_on(&mut self, position: f32) {
+    fn land_on(&mut self, position: i16) {
         self.state_machine = self.state_machine.transition(Event::Land(position));
     }
 
@@ -242,7 +242,7 @@ pub enum Event {
     Slide,
     Jump,
     KnockOut,
-    Land(f32),
+    Land(i16),
     Update,
 }
 
@@ -524,9 +524,9 @@ mod red_hat_boy_states {
             }
         }
 
-        pub fn land_on(self, position: f32) -> RedHatBoyState<Running> {
+        pub fn land_on(self, position: i16) -> RedHatBoyState<Running> {
             RedHatBoyState {
-                context: self.context.set_on(position as i16),
+                context: self.context.set_on(position),
                 _state: Running {},
             }
         }
@@ -563,9 +563,9 @@ mod red_hat_boy_states {
             }
         }
 
-        pub fn land_on(self, position: f32) -> RedHatBoyState<Sliding> {
+        pub fn land_on(self, position: i16) -> RedHatBoyState<Sliding> {
             RedHatBoyState {
-                context: self.context.set_on(position as i16),
+                context: self.context.set_on(position),
                 _state: Sliding {},
             }
         }
@@ -593,9 +593,9 @@ mod red_hat_boy_states {
             }
         }
 
-        pub fn land_on(self, position: f32) -> RedHatBoyState<Running> {
+        pub fn land_on(self, position: i16) -> RedHatBoyState<Running> {
             RedHatBoyState {
-                context: self.context.reset_frame().set_on(position as i16),
+                context: self.context.reset_frame().set_on(position),
                 _state: Running,
             }
         }
@@ -683,8 +683,8 @@ impl Platform {
     }
 
     fn bounding_boxes(&self) -> Vec<Rect> {
-        const X_OFFSET: f32 = 60.0;
-        const END_HEIGHT: f32 = 54.0;
+        const X_OFFSET: i16 = 60;
+        const END_HEIGHT: i16 = 54;
         let destination_box = self.destination_box();
         let bounding_box_one = Rect {
             x: destination_box.x,
@@ -695,7 +695,7 @@ impl Platform {
         let bounding_box_two = Rect {
             x: destination_box.x + X_OFFSET,
             y: destination_box.y,
-            width: destination_box.width - (X_OFFSET * 2.0),
+            width: destination_box.width - (X_OFFSET * 2),
             height: destination_box.height,
         };
         let bounding_box_three = Rect {
